chore: create config file
The configurations were previously stored on the default "funcs.h" file.
We faced a specific recurrent problem: when we changed the configs in
order to run a specific test, make wouldn't detect a change in a
compiled file, and wouldn't recompile the object files. Putting all
object files dependent on "funcs.h" seemed out of place, separating the
configs in a config file and putting all object files dependent on that
one is a nicer solution.